pyproject.toml
[project]
name = "2025-fastapi-seminar-template-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "argon2-cffi>=25.1.0",
    "authlib>=1.6.4",
    "fastapi[standard]>=0.116.1",
    "pyjwt>=2.10.1",
]
